Many things have been added and amended. 1) All fragments and activities are currently reading and writing from the database as should be 2) Many new layouts have been created and editted. 3)New fragments and activities that provided ui functions for searching, creating accounts, logging to accounts have been written.
TODO: Allow user account, textbooks, and request have an attribute added for time! This will be used to order the list and recieve the most recent request (instead of creation id!)
TODO: Create a fragment for password reset and account verification.
TODO: Design the ui (layouts ONLY) to look pretty :)