Parsing book data from google api. Things are working well.
TODO: Either store ro retrieve images without user photos
TODO: Check to see if ISBN or Titles are valid. If not, valid ISBN, tell the user that there is an issue. If invalid Title, change the information. To the correct info.
TODO:Fix image capture from camera. It seems that a valid path is not being created from the camera. Also, it seems that a valid image is not being stored after the image capture has occurred!! FIX THIS